# [IMP] stock: various optimizations when validating inventory.