Backport of 14.0 PR odoo/odoo#73082 changing name to_unlink_candidates -> to_unlink_candidate_ids.

Backport of master PR odoo/odoo#62083

Improve `stock.picking._compute_state` by computing pickings new state conditions
beforehand. This avoids looping through picking.move_lines for each
condition for each picking. Add `picking.move_lines` to defaultdict beforehand
to avoid calling `picking.move_lines` in for loop.

#### Speedup

Validating a 171 lines inventory in a customer DB with 2M stock_moves and 22k stock_pickings. min, max and
avg time shown for each optimization.

Total time taken by `inventory.action_validate()`

| Before PR | Backport 14 only | Backport 14 + Backport master | Backport 14 + Backport master + _compute_state |
|:----------:|:--------------:|:-------------------------------:|:-------------------------------------------------:|
| +1h*| 17min | 13 min | 7min40s |

Avg, Min, Max time taken by `move_line._free_reservation`

| | Before PR | Backport 14 only | Backport 14 + Backport master | Backport 14 + Backport master + _compute_state |
|:--:|:----------:|:--------------:|:-------------------------------:|:-------------------------------------------------:|
| avg | 15s* | 5s | 4.71s | 2s |
| min | 0.001s | 0.001s | 0.001s | 0.001s |
| max | 11min* | 6min | 1min30s | 54s |


_\* validating the inventory 'Before PR' was stopped after 1h while still running. Thus the reported total, avg and max time 'Before PR' is probably underestimated._

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr