@Aurelienvd 2021-08-11
@ryv-odoo Is the PR good as it is now ?